Add serverside rendering (#4)
Moved to Serverside fetching of data and optimised app:

- used `@next/bundle-analyzer` to find main bundle bloat culprits
- loaded several component lazily that would only show up after interaction by importing them with `next/dynamic`
- Manually tree-shook `react-player` (was loading way to much)
- Reimplemented grid layout with gridTemplateColumns in CSS instead of the GridList component of Material UI

Still not too happy with performance for such a simple app.. main culprit of slow loading is now the collection and image backend I'm using "Directus". Collection query is very slow and thumbnails don't have a correct caching policy (which can't be changed).